Barky/requirements.txt
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
pytest==7.2.2
python-dotenv==1.0.0
redis==4.5.2
requests==2.28.2
SQLAlchemy==2.0.7
tenacity==8.2.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
